# swagger.yaml
openapi: 3.0.3
info:
  title: SHTT API - Quản lý Sở hữu Trí tuệ
  description: |
    API quản lý Sở hữu Trí tuệ gồm: Bản quyền tác giả, Nhãn hiệu, Sáng chế,
    Kiểu dáng công nghiệp, Chỉ dẫn địa lý, Làng nghề truyền thống.

    ## Xác thực
    Sử dụng **JWT Bearer Token**. Đăng nhập qua `/api/auth/login` để lấy `accessToken`,
    sau đó nhấn nút **Authorize** và nhập `Bearer <token>`.

    ## Phân quyền
    | Nhóm endpoint | Quyền yêu cầu |
    |---|---|
    | GET /search, GET /, GET /:id | Công khai |
    | GET /search-all | Admin hoặc Staff |
    | POST / PUT / DELETE (IP resources) | Admin hoặc Staff |
    | Mọi thao tác /api/users | Chỉ Admin |

    ## Upload ảnh
    Các endpoint POST/PUT sử dụng `multipart/form-data`.
    Trường `owner`, `authors`, `managementOrg` truyền dưới dạng **JSON string** trong form.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Access Token lấy từ POST /api/auth/login
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Thao tác thành công
    SuccessWithId:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        id:
          type: integer
          format: int64
          example: 1714200000000
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Lỗi server
        error:
          type: string
          example: Error detail

    # ── Stakeholder ─────────────────────────────
    StakeholderBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Nguyễn Văn A
        address:
          type: string
          example: 123 Lê Lợi, Hà Nội
    StakeholderObject:
      allOf:
        - $ref: '#/components/schemas/StakeholderBody'
        - type: object
          properties:
            id:
              type: integer
              format: int64
              example: 1714200000001
...
    CopyrightDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        certificateNumber:
          type: string
          example: BQ-2024-001
        grantDate:
          type: string
          format: date
          example: "2024-01-15"
        title:
          type: string
          example: Phần mềm quản lý nhà hàng
        type:
          type: string
          example: software
        imageUrls:
          type: string
          description: JSON array lưu đường dẫn ảnh
          example: '["/uploads/img1.jpg"]'
        authors:
          type: array
          items:
            $ref: '#/components/schemas/StakeholderObject'
        owners:
          type: array
          items:
            $ref: '#/components/schemas/StakeholderObject'

    # ── Craft Village ────────────────────────────
...
    CopyrightRow:
      type: object
      description: Hàng dữ liệu thô từ bảng copyrights, không có authors/owners lồng nhau
      properties:
        id:
          type: integer
          format: int64
        certificateNumber:
          type: string
          example: BQ-2024-001
        grantDate:
          type: string
          format: date
          example: "2024-01-15"
        title:
          type: string
          example: Phần mềm quản lý nhà hàng
        type:
          type: string
          example: software
        imageUrls:
          type: string
          description: JSON array string chứa đường dẫn ảnh
          example: '["/uploads/img1.jpg"]'
...
  responses:
    Unauthorized:
      description: Chưa xác thực — thiếu hoặc sai token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Unauthorized
    NotFound:
      description: Không tìm thấy tài nguyên
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Không tìm thấy
    ServerError:
      description: Lỗi server nội bộ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Lỗi server
            error: Error detail

# ─────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────
paths:
...
  /api/copyrights:
    get:
      tags: [Copyrights]
      summary: Lấy danh sách tất cả bản quyền
      operationId: getAllCopyrights
      description: Trả về danh sách thô, không có authors/owners lồng nhau. Dùng GET /:id để lấy đầy đủ.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CopyrightRow'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Copyrights]
      summary: Tạo bản quyền mới
      operationId: createCopyright
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                certificateNumber:
                  type: string
                  example: BQ-2024-001
                grantDate:
                  type: string
                  format: date
                  example: "2024-01-15"
                title:
                  type: string
                  example: Phần mềm quản lý nhà hàng
                type:
                  type: string
                  example: software
                authors:
                  type: string
                  description: 'JSON array — [{"name":"...","address":"..."}]'
                  example: '[{"name":"Nguyễn Văn A","address":"123 Lê Lợi, Hà Nội"}]'
                owners:
                  type: string
                  description: 'JSON array — [{"name":"...","address":"..."}]'
                  example: '[{"name":"Công ty TNHH ABC Tech","address":"789 Nguyễn Huệ, TP.HCM"}]'
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: File ảnh, tối đa 5 file
      responses:
        '201':
          description: Tạo thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessWithId'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/copyrights/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1714200000000
    get:
      tags: [Copyrights]
      summary: Xem chi tiết bản quyền
      operationId: getCopyrightById
      description: Trả về bản quyền kèm danh sách `authors` và `owners`.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/CopyrightDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags: [Copyrights]
      summary: Cập nhật bản quyền
      operationId: updateCopyright
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                certificateNumber:
                  type: string
                grantDate:
                  type: string
                  format: date
                title:
                  type: string
                type:
                  type: string
                authors:
                  type: string
                  description: JSON array string
                owners:
                  type: string
                  description: JSON array string
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Bỏ trống để giữ ảnh cũ
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags: [Copyrights]
      summary: Xóa bản quyền
      operationId: deleteCopyright
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Xóa thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  # ════════════════════════════════
  # CRAFT VILLAGES
  # ════════════════════════════════
...